+++ src/Providers/Codeship.php
@@ -5,7 +5,7 @@ namespace MatthiasMullie\CI\Providers;
 use MatthiasMullie\CI\Environment;
 
 /**
- * @see https://codeship.com/documentation/continuous-integration/set-environment-variables/
+ * @see https://codeship.com/documentation/continuous-integration/set-environment-variables
  *
  * @author Matthias Mullie <ci-environment@mullie.eu>
  * @copyright Copyright (c) 2016, Matthias Mullie. All rights reserved.
@@ -35,6 +35,19 @@ class Codeship implements Environment
     public function getRepo()
     {
         return exec('git config --get remote.origin.url');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getSlug()
+    {
+        // only GitHub & BitBucket are supported; both will have repo urls that
+        // include the project slug in this format
+        $url = $this->getRepo();
+        preg_match('/([^:\/]+\/[^:\/]+?)(\.git|$)/', $url, $matches);
+
+        return $matches[1];
     }
 
     /**